==> 2. Scripts/1. Wrangling.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframes to be used  
1: Wrangled data for orders dataframe and departments dataframe    
2: Column datatype change  
3: Rename column  
4: Value counts for order_hour_of_day column   
5: Retrieve department 4 from data dictionary  
6: Subset of products dataframe  
7: Subset of products dataframe  
8: Number of rows and columns from second subset  
9: Table of user_id 1  
10: Profile of user_id 1  
11: Export orders dataframe

# Imports

import pandas
import numpy
import os

# Dataframes

df_orders = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\orders.csv")
df_products = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\products.csv")
df_departments = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\departments.csv")

# 1

# Drop unnecessary eval_set column
df_orders = df_orders.drop(columns = ['eval_set'])

# Rename order_dow column to be more descriptive 
df_orders.rename(columns = {'order_dow':'order_day_of_week'}, inplace = True)

df_orders.head()

# Change column from integer to string
df_orders['order_id'] = df_orders['order_id'].astype('str')

df_orders['order_id'].dtype

# Transpose departments dataframe
df_departments_t = df_departments.T
new_header = df_departments_t.iloc[0]
df_departments_t = df_departments_t[1:]
df_departments_t.columns = new_header

df_departments_t

# Departments data dictionary
data_dict = df_departments_t.to_dict('index')

data_dict

# Subset from products dataframe
df_snacks = df_products[df_products['department_id'] == 19]

df_snacks.head()

# 2

# Change column from integer to string
df_orders['user_id'] = df_orders['user_id'].astype('str')

df_orders.dtypes

# 3

# Rename days_since_prior_order column without overwriting
df_orders.rename(columns = {'days_since_prior_order':'days_since_last_order'})

# 4

df_orders['order_hour_of_day'].value_counts()

Busiest hour of the day for placing orders is the 10th hour

# 5

data_dict.get('4')

department_id 4 represents produce 

# 6

# Subset from products dataframe
df_breakfast = df_products[df_products['department_id'] == 14]

df_breakfast.head()

# 7

# Subset from products dataframe
subset = df_products[df_products['department_id'].isin([7, 12, 20])]

subset.head()

# 8

subset.shape

This subset contains 7650 rows

# 9

# Information about user_id 1
df_orders[df_orders['user_id'] == '1']

# 10

This user orders during the first half of the week. They order at many different hours of the day. They make an order every 2-4 weeks. 

# 11

# Export orders
df_orders.to_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_wrangled.csv")
