==> 2. Scripts/2. Consistency Checks.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframes to be used  
1: Delete missing values and duplicates from products dataframe   
2: Statistics for products dataframe   
3: Check for mixed-type values in orders dataframe   
4: Report for mixed-type data check  
5: Missing values for orders dataframe   
6: Adressing missing values from orders dataframe   
7: Check for duplicates from orders dataframe   
8: Report for duplicates check  
9: Export products dataframe and orders dataframe

# Imports

import pandas
import numpy
import os

# Dataframes

df_orders = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_wrangled.csv")
df_products = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\products.csv")

# 1

# Show number of missing values in each column
df_products.isnull().sum()

# Show all records with a missing product name
df_products[df_products['product_name'].isnull() == True]

# Show number of rows before consistency checks
df_products.shape

# Create new products dataframe without records containing missing values
df_products_consistent = df_products[df_products['product_name'].isnull() == False]
df_products_consistent.shape

# Update dataframe with duplicates removed
df_products_consistent = df_products_consistent.drop_duplicates()
df_products_consistent.shape

# 2

df_products_consistent.describe()

The prices column should be further investigated, because the max is too high.

# 3

# Check for mixed-type values
for col in df_orders.columns.tolist():
  mixtype = (df_orders[[col]].applymap(type) != df_orders[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_orders[mixtype]) > 0:
    print (col)

# 4

No mixed-type data found

# 5

# Check for missing values in the orders dataframe
df_orders.isnull().sum()

There are 206209 missing values in the days_since_prior_order_column. These are probably first-time orders.

# 6

Nothing was done with missing values under the assumption that values are missing because they are first-time orders.

# 7

# Check for duplicates
df_orders[df_orders.duplicated()]

If there were duplicates found, a new dataframe without the duplicates could be created.

# 8

There were no duplicated found.

# 9

# Export dataframes
df_products_consistent.to_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\products_prepared.csv")
df_orders.to_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_prepared.csv")
