==> 2. Scripts/3a. Dataset Merges.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframes to be used  
1: Merge orders dataframe  with orders_prior dataframe   
2: Export merged dataframe

# Imports

import pandas
import numpy
import os

# Dataframes

df_orders_prior = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\order_products_prior.csv")
df_orders = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_prepared.csv")

# 1

# Merge dataframes
df_merged_large = df_orders.merge(df_orders_prior, indicator = True)

df_merged_large['_merge'].value_counts()

# 2

# Export merged dataframe as pkl
df_merged_large.to_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_combined.pkl")
