==> 2. Scripts/3b. Dataset Merges.ipynb <==
### Table of Contents:
Imports: Imported libraries  
1: Dataframe to be used  
2: Number of rows and columns for orders_products_combined dataframe   
3: Merge orders_products_combined dataframe with products dataframe  
4: Check results of merge  
5: Export merged dataframe

# Imports

import pandas
import numpy
import os

# 1

# Import previously combined dataframe as df_orders
df_orders = pandas.read_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_combined.pkl")

# 2

df_orders.shape

# 3

# Import products dataframe
df_products = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\products_prepared.csv")

# Merge dataframes
df_merged = df_orders.merge(df_products, indicator = True)

# 4

# Confirm results of merge
df_merged['_merge'].value_counts()

# 5

# Export merged dataframe
df_merged.to_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")
