==> 2. Scripts/4. Derived Variables.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframe to be used  
1: Derive 2 new columns for orders_products_merged dataframe   
2: New column  
3: Compare derived columns  
4: Derive new column for orders_products_merged dataframe   
5: Value counts for busiest_period_of_day column   
6: Export orders_products_merged dataframe

# Imports

import pandas
import numpy
import os

# Dataframes

orders_products = pandas.read_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")

# 1

# Derive new column price_range
orders_products.loc[orders_products['prices'] <= 5, 'price_range'] = 'Low-range product'

orders_products.loc[(orders_products['prices'] > 5) & (orders_products['prices'] <= 15), 'price_range'] = 'Mid-range product'

orders_products.loc[orders_products['prices'] > 15, 'price_range'] = 'High-range product'

orders_products['price_range'].value_counts()

# Make list to fill new column busiest_day
lst = []
for cell in orders_products['order_day_of_week']:
    if cell == 0:
        lst.append('Busiest day')
    elif cell == 4:
        lst.append('Least busy')
    else:
        lst.append('Regularly busy')

orders_products['busiest_day'] = lst

orders_products['busiest_day'].value_counts()

# 2

# Make list to fill new column busiest_days
lst = []
for cell in orders_products['order_day_of_week']:
    if cell == 0 or cell == 1:
        lst.append('Busiest day')
    elif cell == 4 or cell == 3:
        lst.append('Least busy')
    else:
        lst.append('Regularly busy')

orders_products['busiest_days'] = lst

orders_products['busiest_days'].value_counts()

# 3

The new column busiest_days differs from the column busiest_day in that the value of "Regularly busy" has decreased in occurence and the other two values have increased in occurence. 

# 4

# Make list to fill new column busiest_period_of_day
lst = []
for cell in orders_products['order_hour_of_day']:
    if cell in(10, 11, 14, 15, 13, 12, 16, 9):
        lst.append('Most orders')
    elif cell in(17, 8, 18, 19, 20, 7, 21, 22):
        lst.append('Average Orders')
    else:
        lst.append('Fewest Orders')

orders_products['busiest_period_of_day'] = lst

# 5

# Occurence of values in column busiest_period_of_day
orders_products['busiest_period_of_day'].value_counts()

# 6

# Overwrite dataframe
orders_products.to_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")
