==> 2. Scripts/5. Derived and Aggregated Variables.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframe to be used  
1: Group order_number by department  
2: Comparison of two different groupings  
3: Derive two columns for orders_products_merged dataframe   
4: Statistics for prices of products based on loyalty_status  
5: Derive two columns for orders_products_merged dataframe   
6: Derive two columns for orders_products_merged dataframe   
7: Export dataframe

# Imports

import pandas
import numpy

# Dataframes

orders_products = pandas.read_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")

# 1

# Group average order number by department
orders_products.groupby('department_id')['order_number'].mean()

# 2

Grouping the entire dataframe differs from grouping the subset in that all departments are now included in the output.

# 3

# Make new column total_user_orders
orders_products['total_user_orders'] = orders_products.groupby('user_id')['order_number'].transform(numpy.max)

# Make new column loyalty_status
orders_products.loc[orders_products['total_user_orders'] > 40, 'loyalty_status'] = 'Loyal customer'
orders_products.loc[(orders_products['total_user_orders'] > 10) & (orders_products['total_user_orders'] <= 40), 'loyalty_status'] = 'Regular customer'
orders_products.loc[orders_products['total_user_orders'] <= 10, 'loyalty_status'] = 'New customer'

# Check counts for derived column
orders_products['loyalty_status'].value_counts()

# 4

# Check statistics for price of purchased products
orders_products.groupby('loyalty_status').agg({'prices':['mean', 'min', 'max']})

# 5

# Make new column average_product_price
orders_products['average_product_price'] = orders_products.groupby('user_id')['prices'].transform(numpy.mean)

# Make new column spending_status
orders_products.loc[orders_products['average_product_price'] < 10, 'spending_status'] = 'Low spender'
orders_products.loc[orders_products['average_product_price'] >= 10, 'spending_status'] = 'High spender'

# 6

# Make new column median_days_since_prior_order
orders_products['median_days_since_prior_order'] = orders_products.groupby('user_id')['days_since_prior_order'].transform(numpy.median)

# Make new column order_frequency_status
orders_products.loc[orders_products['median_days_since_prior_order'] > 20, 'order_frequency_status'] = 'Non-frequent customer'
orders_products.loc[(orders_products['median_days_since_prior_order'] > 10) & (orders_products['median_days_since_prior_order'] <= 20), 'order_frequency_status'] = 'Regular customer'
orders_products.loc[orders_products['median_days_since_prior_order'] <= 10, 'order_frequency_status'] = 'Frequent customer'

# 7

# Overwrite dataframe
orders_products.to_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")
