==> 2. Scripts/6. Customers Dataset.ipynb <==
### Table of Contents:
Imports: Imported libraries  
Dataframes: Dataframes to be used  
1: Wrangled data for customers dataframe   
2: Check customers dataframe for missing values and duplicates  
3: Merge customers dataframe with orders_products_merged dataframe   
4: Export merged dataframe

# Imports

import pandas
import numpy

# Dataframes

customers = pandas.read_csv(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Original\customers.csv")
orders_products = pandas.read_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_merged.pkl")

# 1

# Drop unneeded columns
customers.drop(columns = ['First Name', 'Surnam', 'date_joined'], inplace = True)

# Rename columns
customers.rename(columns = {'Gender':'gender', 'STATE':'state', 'Age':'age', 'n_dependants':'number_of_dependants',
                            'fam_status':'marital_status', 'income':'annual_income'}, inplace = True)

# 2

# Check for missing values
customers.isnull().sum()

# Check for mixed data-types
for col in customers.columns.tolist():
  mixtype = (customers[[col]].applymap(type) != customers[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (customers[mixtype]) > 0:
    print (col)

# Check for duplicates
customers[customers.duplicated()]

# 3

# Merge customers dataframe
merged = orders_products.merge(customers, on = 'user_id')

# 4

# Export dataframe
merged.to_pickle(r"C:\Users\justi\Downloads\Instacart Basket Analysis\2. Data\Prepared\orders_products_customers.pkl")
